% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_track_signal_shared.R
\name{.track_all_common_before_fetch}
\alias{.track_all_common_before_fetch}
\title{the idea is that all signal type tracks have a basic framework
1) do stuff before fetch
2) fetch
3) do stuff after fetch}
\usage{
.track_all_common_before_fetch(
  signal_files,
  query_gr,
  fetch_fun = seqsetvis::ssvFetchBam,
  summary_FUN = c("mean", "max")[2],
  flip_x = NULL,
  nwin = 3000,
  nspline = 1,
  nMovingAverage = 1,
  fill_outline_color = NA,
  fill_alpha = 1,
  color_alpha = 1,
  y_label = "signal",
  x_scale = c("bp", "kbp", "Mbp")[2],
  floor_value = 0,
  ceiling_value = Inf,
  color_VAR = NULL,
  color_mapping = NULL,
  fill_VAR = "sample",
  fill_mapping = NULL,
  facet_VAR = "sample",
  legend.position = "right",
  names_on_right = TRUE,
  target_strand = NULL,
  flip_strand = FALSE,
  return_data = FALSE,
  ...
)
}
\description{
.track_all_common_before_fetch
  .track_rna_common_before_fetch
  .track_chip_common_before_fetch
.track_all_common_after_fetch
  .track_rna_common_after_fetch
  .track_chip_common_after_fetch
}
\details{
Used in:
track_rna.PE
track_rna.SE
track_chip.SE
track_chip.PE
}
