% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tracks.R
\name{my_plot_goi}
\alias{my_plot_goi}
\title{my_plot_goi}
\usage{
my_plot_goi(
  goi,
  manual_query_gr = NULL,
  rel_heights.plots = c(3, 1, 3),
  rel_heights.title = c(1, 10),
  peak_grs = NULL,
  nwin = 3000,
  nspline = 10,
  my_query_dt = query_dt,
  show_all_x_axis = FALSE
)
}
\arguments{
\item{show_all_x_axis}{}
}
\description{
my_plot_goi
}
