% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_fetch_bamPE.R
\name{ssvFetchBamPE.RNA_splice}
\alias{ssvFetchBamPE.RNA_splice}
\title{ssvFetchBamPE.RNA_splice}
\usage{
ssvFetchBamPE.RNA_splice(
  file_paths,
  qgr,
  win_size = 50,
  target_strand = "both",
  return_data.table = FALSE,
  win_method = "sample",
  max_dupes = Inf,
  flip_strand = FALSE,
  n_cores = getOption("mc.cores", 1),
  force_skip_centerFix = TRUE,
  n_region_splits = 1,
  ...
)
}
\arguments{
\item{n_region_splits}{}
}
\description{
ssvFetchBamPE.RNA_splice
}
