% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_track_features.R
\name{track_features}
\alias{track_features}
\title{track_features}
\usage{
track_features(
  feature_grs,
  query_gr,
  sample_info_df = NULL,
  sample_info_df.name_VAR = "sample",
  sample_info_df.color_VAR = NULL,
  sample_info_df.fill_VAR = "sample",
  attrib = "cluster_id",
  pad = 0.1,
  flip_x = NULL,
  manual_levels = NULL,
  x_scale = c("bp", "kbp", "Mbp")[2],
  color_mapping = NULL,
  fill_mapping = NULL,
  legend.position = "right"
)
}
\arguments{
\item{manual_levels}{}
}
\description{
track_features
}
