% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_track_ref.R
\name{track_gene_reference}
\alias{track_gene_reference}
\title{track_gene_reference}
\usage{
track_gene_reference(
  ref = "~/../joeboyd/gencode.v36.annotation.gtf",
  query_gr,
  flip_x = FALSE,
  exon_height = 0.4,
  intron_thickness = 2,
  show_tss = FALSE,
  tss_size = 1.5,
  tss_color = "red",
  return_data = FALSE,
  tss_up = 0.4,
  tss_over = 0.05,
  tss_arrow_size = 0.4,
  plus_strand_color = "black",
  minus_strand_color = "darkgray",
  legend.position = "right",
  x_scale = c("bp", "kbp", "Mbp")[2]
)
}
\arguments{
\item{tss_arrow_size}{}
}
\description{
track_gene_reference
}
