% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_track_ref.R
\name{track_gene_transcripts}
\alias{track_gene_transcripts}
\title{track_gene_transcripts}
\usage{
track_gene_transcripts(
  ref = "~/../joeboyd/gencode.v36.annotation.gtf",
  sel_gene_name,
  query_gr = NULL,
  transcript_subset = NULL,
  flip_x = NULL,
  x_scale = c("bp", "kbp", "Mbp")[2]
)
}
\arguments{
\item{x_scale}{}
}
\description{
track_gene_transcripts
}
