% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_track_ref.R
\name{track_ref}
\alias{track_ref}
\title{track_ref}
\usage{
track_ref(
  ref = "~/gencode.v28.annotation.gtf.gz",
  qgr,
  flip_x = FALSE,
  exon_height = 0.4,
  intron_thickness = 2,
  show_tss = FALSE,
  tss_size = 3,
  tss_color = "red",
  return_data = FALSE,
  tss_up = 0.4,
  tss_over = 0.15,
  arrow_size = 0.4
)
}
\arguments{
\item{arrow_size}{}
}
\description{
track_ref
}
