% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_track_rna_SE.R,
%   R/function_track_rna_shared.R
\name{track_rna.SE}
\alias{track_rna.SE}
\title{track_rna.SE}
\usage{
track_rna.SE(
  signal_files,
  query_gr,
  fetch_fun = seqsetvis::ssvFetchBam,
  summary_FUN = c("mean", "max")[2],
  flip_x = NULL,
  nwin = 3000,
  nspline = 1,
  fill_outline_color = NA,
  fill_alpha = 1,
  color_alpha = 1,
  y_label = "signal",
  x_scale = c("bp", "kbp", "Mbp")[2],
  floor_value = 0,
  ceiling_value = Inf,
  color_VAR = NULL,
  color_mapping = NULL,
  fill_VAR = "sample",
  fill_mapping = NULL,
  facet_VAR = "sample",
  legend.position = "right",
  names_on_right = TRUE,
  show_splice = TRUE,
  min_splice_count = 10,
  target_strand = NULL,
  flip_strand = TRUE,
  return_data = FALSE,
  ...
)

track_rna.SE(
  signal_files,
  query_gr,
  fetch_fun = seqsetvis::ssvFetchBam,
  summary_FUN = c("mean", "max")[2],
  flip_x = NULL,
  nwin = 3000,
  nspline = 1,
  fill_outline_color = NA,
  fill_alpha = 1,
  color_alpha = 1,
  y_label = "signal",
  x_scale = c("bp", "kbp", "Mbp")[2],
  floor_value = 0,
  ceiling_value = Inf,
  color_VAR = NULL,
  color_mapping = NULL,
  fill_VAR = "sample",
  fill_mapping = NULL,
  facet_VAR = "sample",
  legend.position = "right",
  names_on_right = TRUE,
  show_splice = TRUE,
  min_splice_count = 10,
  target_strand = NULL,
  flip_strand = TRUE,
  return_data = FALSE,
  ...
)
}
\arguments{
\item{flip_strand}{logical. Should aligned strand be flipped relative to query_gr.  The vast majority of SE libraries are actually flip_strand, so default flip_strand is TRUE.}

\item{...}{}
}
\description{
track_rna.SE

track_rna.SE
}
