% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_asemble_tracks.R
\name{assemble_tracks}
\alias{assemble_tracks}
\title{assemble_tracks}
\usage{
assemble_tracks(plot_list, query_gr, rel_heights = rep(1, length(plot_list)))
}
\arguments{
\item{plot_list}{list of ggplots and/or grobs}

\item{rel_heights}{}
}
\description{
assemble_tracks
}
