% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_fetch_bamPE.R
\name{ssvFetchBamPE.RNA_splice}
\alias{ssvFetchBamPE.RNA_splice}
\title{ssvFetchBamPE.RNA_splice}
\usage{
ssvFetchBamPE.RNA_splice(
  file_paths,
  query_gr,
  win_size = 50,
  target_strand = "both",
  return_data.table = FALSE,
  win_method = "sample",
  max_dupes = Inf,
  flip_strand = FALSE,
  n_cores = getOption("mc.cores", 1),
  force_skip_centerFix = TRUE,
  n_region_splits = 1,
  ...
)
}
\arguments{
\item{file_paths}{character vector of file_paths to load from. Alternatively, file_paths can be a data.frame or data.table whose first column is a character vector of paths and additial columns will be used as metadata.}

\item{query_gr}{Set of GRanges to query. For valid results the width of each interval should be identical and evenly divisible by win_size.}

\item{win_size}{The window size that evenly divides widths in qgr.}

\item{target_strand}{character. if one of "+" or "-", reads are filtered to match. ignored if any other value.}

\item{return_data.table}{logical. If TRUE the internal data.table is returned instead of GRanges. Default is FALSE.}

\item{win_method}{character. one of c("sample", "summary"). "sample" selects values at intervals and "summary" applies a weight mean function to all values in window.}

\item{max_dupes}{numeric >= 1. duplicate reads by strandd start position over this number are removed, Default is Inf.}

\item{flip_strand}{logical. if TRUE strands are flipped.}

\item{n_cores}{integer number of cores to use. Uses mc.cores option if not supplied.}

\item{force_skip_centerFix}{boolean, if TRUE all query ranges will be used "as is". This is already the case by default if win_method == "summary" but may have applications where win_method == "sample".}

\item{n_region_splits}{integer number of splits to apply to qgr. The query GRanges will be split into this many roughly equal parts for increased parallelization. Default is 1, no split.}

\item{sum_reads}{logical. If true R1 and R2 reads are added together. If FALSE they are returned separately, identified by the "read" attribute.}
}
\value{
A tidy formatted GRanges (or data.table if specified) containing fetched values.
}
\description{
ssvFetchBamPE.RNA_splice
}
\examples{
pkg_dir = system.file(package = "ssvTracks", "extdata", mustWork = TRUE)
bam_files_runx1 = dir(pkg_dir, pattern = "RUNX1_RNA.+bam$", full.names = TRUE)
names(bam_files_runx1) = sub("_rep.+", "", basename(bam_files_runx1))
bed_file_runx1 = dir(pkg_dir, pattern = "RUNX1.bed", full.names = TRUE)
query_gr = rtracklayer::import.bed(bed_file_runx1)

track_rna.PE(bam_files_runx1, query_gr)
track_rna.PE(bam_files_runx1, query_gr, flip_strand = TRUE)

splice_dt.no_flip = ssvFetchBamPE.RNA_splice(bam_files_runx1, query_gr, return_data.table = TRUE)
splice_dt.no_flip
splice_dt = ssvFetchBamPE.RNA_splice(bam_files_runx1, query_gr, return_data.table = TRUE, flip_strand = TRUE)
splice_dt

}
