% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_track_chip.R
\name{track_chip}
\alias{track_chip}
\title{track_chip}
\usage{
track_chip(
  bw_files,
  qgr,
  fetch_fun = ssvFetchBam,
  win_FUN = c("mean", "max")[2],
  sum_FUN = NULL,
  flip_x = NULL,
  nwin = 3000,
  nspline = 10,
  show_lines = FALSE,
  show_fill = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
track_chip
}
