% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_track_features.R
\name{track_features.numeric}
\alias{track_features.numeric}
\title{track_features.numeric}
\usage{
track_features.numeric(
  feature_grs,
  query_gr,
  color_VAR = NULL,
  fill_VAR = NULL,
  attrib = "cluster_id",
  pad = 0.1,
  flip_x = NULL,
  manual_levels = NULL,
  x_scale = c("bp", "kbp", "Mbp")[2],
  color_mapping = NULL,
  fill_mapping = NULL,
  legend.position = "right"
)
}
\arguments{
\item{legend.position}{}
}
\description{
track_features.numeric
}
