% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_track_features.R
\name{track_features.numeric}
\alias{track_features.numeric}
\title{track_features.numeric}
\usage{
track_features.numeric(
  feature_grs,
  query_gr,
  color_VAR = NULL,
  fill_VAR = NULL,
  attrib = "cluster_id",
  pad = 0.1,
  flip_x = NULL,
  manual_levels = NULL,
  x_scale = c("bp", "kbp", "Mbp")[2],
  color_mapping = NULL,
  fill_mapping = NULL,
  legend.position = "right"
)
}
\arguments{
\item{legend.position}{}
}
\description{
track_features.numeric
}
\examples{

clust_dt = seqsetvis::ssvSignalClustering(seqsetvis::CTCF_in_10a_profiles_dt)
assign_dt = unique(clust_dt[, .(id, cluster_id)])
olap_gr = seqsetvis::CTCF_in_10a_overlaps_gr
olap_gr$cluster_id = ""
olap_gr[assign_dt$id]$cluster_id = assign_dt$cluster_id
track_features.numeric(olap_gr,
                       query_gr = GRanges("chr1", IRanges(40e6, 70e6)),
                       color_VAR = "cluster_id",
                       fill_VAR = "cluster_id",
                       attrib = "peaks")
}
