% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_track_features.R
\name{track_features}
\alias{track_features}
\title{track_features}
\usage{
track_features(
  feature_grs,
  query_gr,
  attrib = "cluster_id",
  pad = 0.1,
  flip_x = NULL,
  manual_levels = NULL
)
}
\arguments{
\item{manual_levels}{}
}
\description{
track_features
}
