% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_track_rna_PE.R,
%   R/function_track_rna_shared.R
\name{track_rna.PE}
\alias{track_rna.PE}
\title{track_rna.PE}
\usage{
track_rna.PE(
  signal_files,
  query_gr,
  summary_FUN = c("mean", "max")[2],
  flip_x = NULL,
  nwin = 3000,
  nspline = 1,
  fill_outline_color = NA,
  fill_alpha = 1,
  color_alpha = 1,
  y_label = "signal",
  x_scale = c("bp", "kbp", "Mbp")[2],
  floor_value = 0,
  ceiling_value = Inf,
  color_VAR = NULL,
  color_mapping = NULL,
  fill_VAR = "sample",
  fill_mapping = NULL,
  facet_VAR = "sample",
  legend.position = "right",
  names_on_right = TRUE,
  show_splice = TRUE,
  min_splice_count = 10,
  target_strand = NULL,
  flip_strand = FALSE,
  return_data = FALSE,
  ...
)

track_rna.PE(
  signal_files,
  query_gr,
  summary_FUN = c("mean", "max")[2],
  flip_x = NULL,
  nwin = 3000,
  nspline = 1,
  fill_outline_color = NA,
  fill_alpha = 1,
  color_alpha = 1,
  y_label = "signal",
  x_scale = c("bp", "kbp", "Mbp")[2],
  floor_value = 0,
  ceiling_value = Inf,
  color_VAR = NULL,
  color_mapping = NULL,
  fill_VAR = "sample",
  fill_mapping = NULL,
  facet_VAR = "sample",
  legend.position = "right",
  names_on_right = TRUE,
  show_splice = TRUE,
  min_splice_count = 10,
  target_strand = NULL,
  flip_strand = FALSE,
  return_data = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
track_rna.PE

track_rna.PE
}
