% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_track_rna_SE.R
\name{track_rna.SE}
\alias{track_rna.SE}
\title{track_rna.SE}
\usage{
track_rna.SE(
  signal_files,
  query_gr,
  fetch_fun = seqsetvis::ssvFetchBam,
  summary_FUN = c("mean", "max")[2],
  flip_x = NULL,
  nwin = 3000,
  nspline = 1,
  fill_outline_color = NA,
  fill_alpha = 1,
  color_alpha = 1,
  y_label = "signal",
  x_scale = c("bp", "kbp", "Mbp")[2],
  floor_value = 0,
  ceiling_value = Inf,
  color_VAR = NULL,
  color_mapping = NULL,
  fill_VAR = "sample",
  fill_mapping = NULL,
  facet_VAR = "sample",
  legend.position = "right",
  names_on_right = TRUE,
  show_splice = TRUE,
  min_splice_count = 0,
  splice_within_range_only = NULL,
  target_strand = NULL,
  flip_strand = TRUE,
  return_data = FALSE,
  ...
)
}
\arguments{
\item{signal_files}{Character or data.frame. Paths to files. Either supplied as a simple character vector (names are use for plot labels if set) or as a data.frame}

\item{fetch_fun}{An ssvFetc* function from seqsetvis. ssvFetchBam, ssvFetchBamPE or ssvFetchBigwig are likely choices.}

\item{summary_FUN}{Either a function or character "mean" or "max".  If a custom function it must follow the form of weighted.mean() and accept 2 arguments, values and weights.}

\item{nwin}{Numeric. Higher numbers increase resolution but increase plotting time and size of .pdf files.}

\item{nspline}{Numeric. If higher than 1, splines will be used to interpolate and smooth between windows.}

\item{fill_outline_color}{Character. Color applied to outline for geom_ribbon used for filled tracks.}

\item{y_label}{Character. Label for y-axis.}

\item{floor_value}{Numeric.  Values below floor will be increased to floor. Default is 0.}

\item{ceiling_value}{Numeric. Values above ceiling will be decreased to ceiling. Default is Inf.}

\item{color_VAR}{Character. Color variable if supplying signal_files as data.frame. Default of NULL disables colored line.}

\item{color_mapping}{Named character. Maps values of color_VAR to valid colors.}

\item{fill_VAR}{Character. Fill variable if supplying signal_files as data.frame. Default of "sample" results in 1 fill color per file.}

\item{fill_mapping}{Named character. Maps values of fill_VAR to valid colors.}

\item{facet_VAR}{Character.  Files that share a facet value will appear in the same track row. Default of "sample" results in 1 file per row.}

\item{legend.position}{Charactter. Position for legend, see ggplot2::theme.  Most likely "right", "bottom", or "none".}

\item{names_on_right}{logical. If TRUE (default) facet/row names appear on the right. If FALSE facet/row names appear on the left.}

\item{...}{Currently not used.}
}
\description{
track_rna.SE
}
\examples{
pkg_dir = system.file(package = "ssvTracks", "extdata", mustWork = TRUE)
bam_files_esr1 = dir(pkg_dir, pattern = "M.+R1.ESR1_RNA.+bam$", full.names = TRUE)
names(bam_files_esr1) = sub("_R.+", "", basename(bam_files_esr1))
bed_file_esr1 = dir(pkg_dir, pattern = "ESR1.bed", full.names = TRUE)
query_gr = rtracklayer::import.bed(bed_file_esr1)

track_rna.SE(bam_files_esr1, query_gr)

track_rna.SE(bam_files_esr1,
  query_gr,
  show_splice = TRUE,
  flip_strand = TRUE)

track_rna.SE(bam_files_esr1,
  query_gr,
  show_splice = TRUE,
  flip_strand = FALSE)

track_rna.SE(bam_files_esr1,
  GRanges("chr6", IRanges(151800000, 151900000)),
  flip_strand = TRUE,
  show_splice = TRUE)

track_rna.SE(bam_files_esr1,
  GRanges("chr6", IRanges(151800000, 151900000)),
  flip_strand = TRUE,
  show_splice = TRUE,
  color_VAR = "sample")
}
